backend/requirements.txt
Flask==2.0.2
Flask-CORS==3.0.10
Flask-Marshmallow==0.14.0
Flask-Restless-NG==2.2.9
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.26.1
python-dotenv==0.19.1
uWSGI==2.0.19.1
tqdm==4.62.2
requests==2.21.0
SQLAlchemy==1.4.26
psycopg2-binary==2.9.1